SXShop/requirements.txt
django
pillow
mysqlclient
djangorestframework
markdown
django-filter
django-reversion
django-formtools
future
httplib2
six
requests
coreapi
django-cors-headers
xlwt
xlsxwriter
drf-extensions
django-redis
pycryptodome
raven
djangorestframework-jwt
django-crispy-forms